requirements.txt
asgiref==3.3.2
autopep8==1.5
Django==3.2
django-admin-autocomplete-filter==0.4
django-admin-shortcuts==2.0.0
django-autocompletefilter==0.0.6
django-currentuser==0.5.3
django-extensions==2.2.8
django-fontawesome==1.0
django-rest-framework==0.1.0
django-shell-plus==1.1.7
djangorestframework==3.11.0
pycodestyle==2.5.0
pytz==2019.3
PyYAML==5.3
six==1.14.0
sqlparse==0.3.0
django-cors-headers==3.14.0
factory-boy==3.2.1
django-filter
ipdb